Feat-결제추가 (창수)

+++ namu/src/main/java/com/forest/namu/domain/SessionInfo.java
@@ -5,17 +5,17 @@ public class SessionInfo {
 	private long memberIdx;
 	private String userId;
 	private String nickName;
-	private long point;
+	private int point;
 	private int membership;
 	private int enabled;
 	private String town1;
 	private String town2;
 	
 	
-	public long getPoint() {
+	public int getPoint() {
 		return point;
 	}
-	public void setPoint(long point) {
+	public void setPoint(int point) {
 		this.point = point;
 	}
 	public String getTown1() {
